Set up pageinfo middleware for reuse (#54641)
