+++ src/article-api/middleware/pageinfo.ts
@@ -1,5 +1,6 @@
 import express from 'express'
-import type { NextFunction, RequestHandler, Response } from 'express'
+import type { RequestHandler, Response } from 'express'
+import type { ExtendedRequestWithPageInfo } from '../types'
 
 import type { ExtendedRequest, Page, Context, Permalink } from '@/types'
 import statsd from '@/observability/lib/statsd.js'
@@ -13,9 +14,8 @@ import {
 import shortVersions from '@/versions/middleware/short-versions.js'
 import contextualize from '@/frame/middleware/context/context'
 import features from '@/versions/middleware/features.js'
-import getRedirect from '@/redirects/lib/get-redirect.js'
-import { isArchivedVersionByPath } from '@/archives/lib/is-archived-version'
 import { readCompressedJsonFile } from '@/frame/lib/read-json-file.js'
+import { pathValidationMiddleware, pageValidationMiddleware } from './validation'
 
 const router = express.Router()
 
@@ -26,98 +26,6 @@ const router = express.Router()
 // Note! The only reason this variable is exported is so that
 // it can be imported by the script scripts/precompute-pageinfo.ts
 export const CACHE_FILE_PATH = '.pageinfo-cache.json.br'
-
-type ArchivedVersion = {
-  isArchived?: boolean
-  requestedVersion?: string
-}
-
-type ExtendedRequestWithPageInfo = ExtendedRequest & {
-  pageinfo: {
-    pathname: string
-    page?: Page
-    archived?: ArchivedVersion
-  }
-}
-
-const validationMiddleware = (
-  req: ExtendedRequestWithPageInfo,
-  res: Response,
-  next: NextFunction,
-) => {
-  const pathname = req.query.pathname as string | string[] | undefined
-  if (!pathname) {
-    return res.status(400).json({ error: `No 'pathname' query` })
-  }
-  if (Array.isArray(pathname)) {
-    return res.status(400).json({ error: "Multiple 'pathname' keys" })
-  }
-  if (!pathname.trim()) {
-    return res.status(400).json({ error: `'pathname' query empty` })
-  }
-  if (!pathname.startsWith('/')) {
-    return res.status(400).json({ error: `'pathname' has to start with /` })
-  }
-  if (/\s/.test(pathname)) {
-    return res.status(400).json({ error: `'pathname' cannot contain whitespace` })
-  }
-  req.pageinfo = { pathname }
-  return next()
-}
-
-const pageinfoMiddleware = (
-  req: ExtendedRequestWithPageInfo,
-  res: Response,
-  next: NextFunction,
-) => {
-  let { pathname } = req.pageinfo
-  // We can't use the `findPage` middleware utility function because we
-  // need to know when the pathname is a redirect.
-  // This is important so that the final `pathname` value
-  // matches the page's permalinks.
-  // This is important when rendering a page because of translations,
-  // if it needs to do a fallback, it needs to know the correct
-  // equivalent English page.
-
-  if (!req.context || !req.context.pages || !req.context.redirects)
-    throw new Error('request not yet contextualized')
-
-  const redirectsContext = { pages: req.context.pages, redirects: req.context.redirects }
-
-  // Similar to how the `handle-redirects.js` middleware works, let's first
-  // check if the URL is just having a trailing slash.
-  while (pathname.endsWith('/') && pathname.length > 1) {
-    pathname = pathname.slice(0, -1)
-  }
-
-  // E.g. a request for `/` is handled as a redirect outside the
-  // getRedirect() function.
-  if (pathname === '/') {
-    pathname = `/${req.context.currentLanguage}`
-  }
-
-  if (!(pathname in req.context.pages)) {
-    // If a pathname is not a known page, it might *either* be a redirect,
-    // or an archived enterprise version, or both.
-    // That's why it's import to not bother looking at the redirects
-    // if the pathname is an archived enterprise version.
-    // This mimics how our middleware work and their order.
-    req.pageinfo.archived = isArchivedVersionByPath(pathname) as ArchivedVersion
-    if (!req.pageinfo.archived.isArchived) {
-      const redirect = getRedirect(pathname, redirectsContext)
-      if (redirect) {
-        pathname = redirect
-      }
-    }
-  }
-
-  // Remember this might yield undefined if the pathname is not a page
-  req.pageinfo.page = req.context.pages[pathname]
-  // The pathname might have changed if it was a redirect
-  req.pageinfo.pathname = pathname
-
-  return next()
-}
 
 export async function getPageInfo(page: Page, pathname: string) {
   const mockedContext: Context = {}
@@ -226,8 +134,8 @@ async function getPageInfoFromCache(page: Page, pathname: string) {
 
 router.get(
   '/v1',
-  validationMiddleware as RequestHandler,
-  pageinfoMiddleware as RequestHandler,
+  pathValidationMiddleware as RequestHandler,
+  pageValidationMiddleware as RequestHandler,
   catchMiddlewareError(async function pageInfo(req: ExtendedRequestWithPageInfo, res: Response) {
     // Remember, the `validationMiddleware` will use redirects if the
     // `pathname` used is a redirect (e.g. /en/articles/foo or
