+++ src/article-api/types.ts
@@ -1,0 +1,14 @@
+import type { ExtendedRequest, Page } from '@/types'
+
+export type ArchivedVersion = {
+  isArchived?: boolean
+  requestedVersion?: string
+}
+
+export type ExtendedRequestWithPageInfo = ExtendedRequest & {
+  pageinfo: {
+    pathname: string
+    page?: Page
+    archived?: ArchivedVersion
+  }
+}
